+++ modules/estimate.mod.sh
@@ -15,11 +15,16 @@
 
 # MOD ==========================================================================
 
+source "`dirname ${BASH_SOURCE}`/functions.mod.sh"
+
 # Estimation -------------------------------------------------------------------
 echo -e " Estimating centrality ..."
 
 # Prepare paste string
 spaste=""; for i in $(seq 1 ${#bedfiles[@]}); do spaste="$spaste -"; done
+
+# Prepare output table
+metrics=$(echo -e "$comb" | cut -f1-3 | uniq)
 
 # Probability metrics ----------------------------------------------------------
 
@@ -27,88 +32,133 @@ spaste=""; for i in $(seq 1 ${#bedfiles[@]}); do spaste="$spaste -"; done
 echo -e " > Probability ..."
 prob_mat=$(echo -e "$comb" | cut -f4,5,8 | \
     gawk -v type="p" -f "$awkdir/pre_process.awk" | paste $spaste)
-probability_two_points=$(echo -e "$prob_mat" | \
-    gawk -v calc="ratio" -v type="2p" -f "$awkdir/estimate_centrality.awk")
-probability_fixed=$(echo -e "$prob_mat" | \
-    gawk -v calc="ratio" -v type="f" -f "$awkdir/estimate_centrality.awk")
-probability_global=$(echo -e "$prob_mat" | \
-    gawk -v calc="ratio" -v type="g" -f "$awkdir/estimate_centrality.awk")
+
+containsElement "prob_2p" "${calc_metrics[@]}"; if [[ 1 == $? ]]; then
+    probability_two_points=$(echo -e "$prob_mat" | \
+        gawk -v calc="ratio" -v type="2p" -f "$awkdir/estimate_centrality.awk")
+    metrics=$(echo -e "$metrics" | paste -d$'\t' - \
+        <(echo -e "$probability_two_points"))
+fi
+containsElement "prob_f" "${calc_metrics[@]}"; if [[ 1 == $? ]]; then
+    probability_fixed=$(echo -e "$prob_mat" | \
+        gawk -v calc="ratio" -v type="f" -f "$awkdir/estimate_centrality.awk")
+    metrics=$(echo -e "$metrics" | paste -d$'\t' - \
+        <(echo -e "$probability_fixed"))
+fi
+containsElement "prob_g" "${calc_metrics[@]}"; if [[ 1 == $? ]]; then
+    probability_global=$(echo -e "$prob_mat" | \
+        gawk -v calc="ratio" -v type="g" -f "$awkdir/estimate_centrality.awk")
+    metrics=$(echo -e "$metrics" | paste -d$'\t' - \
+        <(echo -e "$probability_global"))
+fi
 
 # Cumulative ratio metric
 echo -e " > Cumulative ratio ..."
 cumrat_mat="$prob_mat"
-cumrat_two_points=$(echo -e "$cumrat_mat" | \
-    gawk -v calc="ratio" -v cumrat=1 -v type="2p" \
-    -f "$awkdir/estimate_centrality.awk")
-cumrat_fixed=$(echo -e "$cumrat_mat" | \
-    gawk -v calc="ratio" -v cumrat=1 -v type="f" \
-    -f "$awkdir/estimate_centrality.awk")
-cumrat_global=$(echo -e "$cumrat_mat" | \
-    gawk -v calc="ratio" -v cumrat=1 -v type="g" \
-    -f "$awkdir/estimate_centrality.awk")
+
+containsElement "cor_2p" "${calc_metrics[@]}"; if [[ 1 == $? ]]; then
+    cumrat_two_points=$(echo -e "$cumrat_mat" | \
+        gawk -v calc="ratio" -v cumrat=1 -v type="2p" \
+        -f "$awkdir/estimate_centrality.awk")
+    metrics=$(echo -e "$metrics" | paste -d$'\t' - \
+        <(echo -e "$cumrat_two_points"))
+fi
+containsElement "cor_f" "${calc_metrics[@]}"; if [[ 1 == $? ]]; then
+    cumrat_fixed=$(echo -e "$cumrat_mat" | \
+        gawk -v calc="ratio" -v cumrat=1 -v type="f" \
+        -f "$awkdir/estimate_centrality.awk")
+    metrics=$(echo -e "$metrics" | paste -d$'\t' - \
+        <(echo -e "$cumrat_fixed"))
+fi
+containsElement "cor_g" "${calc_metrics[@]}"; if [[ 1 == $? ]]; then
+    cumrat_global=$(echo -e "$cumrat_mat" | \
+        gawk -v calc="ratio" -v cumrat=1 -v type="g" \
+        -f "$awkdir/estimate_centrality.awk")
+    metrics=$(echo -e "$metrics" | paste -d$'\t' - \
+        <(echo -e "$cumrat_global"))
+fi
 
 # Ratio cumulative metric
 echo -e " > Ratio cumulative ..."
 ratcum_mat=$(echo -e "$comb" | cut -f4,5,8 | \
     tr '\t' ',' | paste $spaste)
-ratcum_two_points=$(echo -e "$ratcum_mat" | \
-    gawk -v calc="ratio" -v ratcum=1 -v type="2p" \
-    -f "$awkdir/estimate_centrality.awk")
-ratcum_fixed=$(echo -e "$ratcum_mat" | \
-    gawk -v calc="ratio" -v ratcum=1 -v type="f" \
-    -f "$awkdir/estimate_centrality.awk")
-ratcum_global=$(echo -e "$ratcum_mat" | \
-    gawk -v calc="ratio" -v ratcum=1 -v type="g" \
-    -f "$awkdir/estimate_centrality.awk")
+
+containsElement "roc_2p" "${calc_metrics[@]}"; if [[ 1 == $? ]]; then
+    ratcum_two_points=$(echo -e "$ratcum_mat" | \
+        gawk -v calc="ratio" -v ratcum=1 -v type="2p" \
+        -f "$awkdir/estimate_centrality.awk")
+    metrics=$(echo -e "$metrics" | paste -d$'\t' - \
+        <(echo -e "$ratcum_two_points"))
+fi
+containsElement "roc_f" "${calc_metrics[@]}"; if [[ 1 == $? ]]; then
+    ratcum_fixed=$(echo -e "$ratcum_mat" | \
+        gawk -v calc="ratio" -v ratcum=1 -v type="f" \
+        -f "$awkdir/estimate_centrality.awk")
+    metrics=$(echo -e "$metrics" | paste -d$'\t' - \
+        <(echo -e "$ratcum_fixed"))
+fi
+containsElement "roc_g" "${calc_metrics[@]}"; if [[ 1 == $? ]]; then
+    ratcum_global=$(echo -e "$ratcum_mat" | \
+        gawk -v calc="ratio" -v ratcum=1 -v type="g" \
+        -f "$awkdir/estimate_centrality.awk")
+    metrics=$(echo -e "$metrics" | paste -d$'\t' - \
+        <(echo -e "$ratcum_global"))
+fi
 
 # Variability metrics ----------------------------------------------------------
 
 # Variance metric
 echo -e " > Variance ..."
 var_mat=$(echo -e "$comb" | cut -f7 | paste $spaste)
-var_two_points=$(echo -e "$var_mat" | \
-    gawk -v calc="logratio" -v type="2p" -f "$awkdir/estimate_centrality.awk")
-var_fixed=$(echo -e "$var_mat" | \
-    gawk -v calc="logratio" -v type="f" -f "$awkdir/estimate_centrality.awk")
+
+containsElement "var_2p" "${calc_metrics[@]}"; if [[ 1 == $? ]]; then
+    var_two_points=$(echo -e "$var_mat" | \
+        gawk -v calc="logratio" -v type="2p" -f "$awkdir/estimate_centrality.awk")
+    metrics=$(echo -e "$metrics" | paste -d$'\t' - \
+        <(echo -e "$var_two_points"))
+fi
+containsElement "var_f" "${calc_metrics[@]}"; if [[ 1 == $? ]]; then
+    var_fixed=$(echo -e "$var_mat" | \
+        gawk -v calc="logratio" -v type="f" -f "$awkdir/estimate_centrality.awk")
+    metrics=$(echo -e "$metrics" | paste -d$'\t' - \
+        <(echo -e "$var_fixed"))
+fi
 
 # Fano factor metric
 echo -e " > Fano factor ..."
 ff_mat=$(echo -e "$comb" | cut -f6,7 | \
     gawk -v type="ff" -f "$awkdir/pre_process.awk" | paste $spaste)
-ff_two_points=$(echo -e "$ff_mat" | \
-    gawk -v calc="diff" -v type="2p" -f "$awkdir/estimate_centrality.awk")
-ff_fixed=$(echo -e "$ff_mat" | \
-    gawk -v calc="diff" -v type="f" -f "$awkdir/estimate_centrality.awk")
+
+containsElement "ff_2p" "${calc_metrics[@]}"; if [[ 1 == $? ]]; then
+    ff_two_points=$(echo -e "$ff_mat" | \
+        gawk -v calc="diff" -v type="2p" -f "$awkdir/estimate_centrality.awk")
+    metrics=$(echo -e "$metrics" | paste -d$'\t' - \
+        <(echo -e "$ff_two_points"))
+fi
+containsElement "ff_f" "${calc_metrics[@]}"; if [[ 1 == $? ]]; then
+    ff_fixed=$(echo -e "$ff_mat" | \
+        gawk -v calc="diff" -v type="f" -f "$awkdir/estimate_centrality.awk")
+    metrics=$(echo -e "$metrics" | paste -d$'\t' - \
+        <(echo -e "$ff_fixed"))
+fi
 
 # Coefficient of variation metric
 echo -e " > Coefficient of variation ..."
 cv_mat=$(echo -e "$comb" | cut -f6,7 | \
     gawk -v type="cv" -f "$awkdir/pre_process.awk" | paste $spaste)
-cv_two_points=$(echo -e "$cv_mat" | \
-    gawk -v calc="diff" -v type="2p" -f "$awkdir/estimate_centrality.awk")
-cv_fixed=$(echo -e "$cv_mat" | \
-    gawk -v calc="diff" -v type="f" -f "$awkdir/estimate_centrality.awk")
 
-# Assemble ---------------------------------------------------------------------
-
-# Prepare output table
-metrics=$(echo -e "$comb" | cut -f1-3 | uniq | paste -d$'\t' - \
-    <(echo -e "$probability_two_points") \
-    <(echo -e "$probability_fixed") \
-    <(echo -e "$probability_global") \
-    <(echo -e "$cumrat_two_points") \
-    <(echo -e "$cumrat_fixed") \
-    <(echo -e "$cumrat_global") \
-    <(echo -e "$ratcum_two_points") \
-    <(echo -e "$ratcum_fixed") \
-    <(echo -e "$ratcum_global") \
-    <(echo -e "$var_two_points") \
-    <(echo -e "$var_fixed") \
-    <(echo -e "$ff_two_points") \
-    <(echo -e "$ff_fixed") \
-    <(echo -e "$cv_two_points") \
-    <(echo -e "$cv_fixed") \
-    )
+containsElement "cv_2p" "${calc_metrics[@]}"; if [[ 1 == $? ]]; then
+    cv_two_points=$(echo -e "$cv_mat" | \
+        gawk -v calc="diff" -v type="2p" -f "$awkdir/estimate_centrality.awk")
+    metrics=$(echo -e "$metrics" | paste -d$'\t' - \
+        <(echo -e "$cv_two_points"))
+fi
+containsElement "cv_f" "${calc_metrics[@]}"; if [[ 1 == $? ]]; then
+    cv_fixed=$(echo -e "$cv_mat" | \
+        gawk -v calc="diff" -v type="f" -f "$awkdir/estimate_centrality.awk")
+    metrics=$(echo -e "$metrics" | paste -d$'\t' - \
+        <(echo -e "$cv_fixed"))
+fi
 
 # END ==========================================================================